moved edit* to separate dir
