+++ mscore/edit/edit.cmake
@@ -1,0 +1,35 @@
+
+
+set (MSCORE_EDIT_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/editdrumset.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editdrumset.h
+    ${CMAKE_CURRENT_LIST_DIR}/editelement.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editfiguredbass.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editharmony.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editinstrument.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editinstrument.h
+    ${CMAKE_CURRENT_LIST_DIR}/editlyrics.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editpitch.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editpitch.h
+    ${CMAKE_CURRENT_LIST_DIR}/editraster.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editraster.h
+    ${CMAKE_CURRENT_LIST_DIR}/editstaff.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editstaff.h
+    ${CMAKE_CURRENT_LIST_DIR}/editstafftype.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editstafftype.h
+    ${CMAKE_CURRENT_LIST_DIR}/editstringdata.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editstringdata.h
+    ${CMAKE_CURRENT_LIST_DIR}/editstyle.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/editstyle.h
+    )
+
+set (MSCORE_EDIT_UI
+    ${CMAKE_CURRENT_LIST_DIR}/editdrumset.ui
+    ${CMAKE_CURRENT_LIST_DIR}/editinstrument.ui
+    ${CMAKE_CURRENT_LIST_DIR}/editpitch.ui
+    ${CMAKE_CURRENT_LIST_DIR}/editraster.ui
+    ${CMAKE_CURRENT_LIST_DIR}/editstafftype.ui
+    ${CMAKE_CURRENT_LIST_DIR}/editstaff.ui
+    ${CMAKE_CURRENT_LIST_DIR}/editstringdata.ui
+    ${CMAKE_CURRENT_LIST_DIR}/editstyle.ui
+    )
