+++ mscore/edit/editdrumset.ui
