+++ mscore/edit/editelement.cpp
