+++ mscore/edit/editfiguredbass.cpp
