+++ mscore/edit/editharmony.cpp
