+++ mscore/edit/editinstrument.cpp
