+++ mscore/edit/editinstrument.h
