+++ mscore/edit/editinstrument.ui
