+++ mscore/edit/editlyrics.cpp
