+++ mscore/edit/editpitch.cpp
