+++ mscore/edit/editpitch.h
