+++ mscore/edit/editpitch.ui
