+++ mscore/edit/editraster.cpp
