+++ mscore/edit/editraster.h
