+++ mscore/edit/editstaff.cpp
