+++ mscore/edit/editstaff.h
