+++ mscore/edit/editstaff.ui
