+++ mscore/edit/editstafftype.cpp
