+++ mscore/edit/editstafftype.h
