+++ mscore/edit/editstafftype.ui
