+++ mscore/edit/editstringdata.cpp
