+++ mscore/edit/editstringdata.h
