+++ mscore/edit/editstringdata.ui
