+++ mscore/edit/editstyle.cpp
