+++ mscore/edit/editstyle.h
