+++ mscore/edit/editstyle.ui
