+++ mscore/musescore.cpp
@@ -31,7 +31,7 @@
 #include "libmscore/sym.h"
 #include "pagesettings.h"
 #include "debugger/debugger.h"
-#include "editstyle.h"
+#include "edit/editstyle.h"
 #include "playpanel.h"
 #include "libmscore/page.h"
 #include "mixer/mixer.h"
@@ -70,7 +70,7 @@
 #include "libmscore/scorediff.h"
 #include "libmscore/chord.h"
 #include "libmscore/segment.h"
-#include "editraster.h"
+#include "edit/editraster.h"
 #include "pianotools.h"
 #include "mediadialog.h"
 #include "workspace.h"
@@ -91,7 +91,7 @@
 #endif
 #include "helpBrowser.h"
 #include "drumtools.h"
-#include "editstafftype.h"
+#include "edit/editstafftype.h"
 #include "texttools.h"
 #include "textpalette.h"
 #include "resourceManager.h"
